+++ tfe/provider_test.go
@@ -369,12 +369,21 @@ func testAccGithubPreCheck(t *testing.T) {
 	}
 }
 
+func testAccGHAInstallationPreCheck(t *testing.T) {
+	testAccPreCheck(t)
+	if envGithubAppInstallationID == "" {
+		t.Skip("Please set GITHUB_APP_INSTALLATION_ID to run this test")
+	}
+}
+
 func init() {
 	envGithubPolicySetIdentifier = os.Getenv("GITHUB_POLICY_SET_IDENTIFIER")
 	envGithubPolicySetBranch = os.Getenv("GITHUB_POLICY_SET_BRANCH")
 	envGithubPolicySetPath = os.Getenv("GITHUB_POLICY_SET_PATH")
 	envGithubRegistryModuleIdentifer = os.Getenv("GITHUB_REGISTRY_MODULE_IDENTIFIER")
 	envGithubToken = os.Getenv("GITHUB_TOKEN")
+	envGithubAppInstallationID = os.Getenv("GITHUB_APP_INSTALLATION_ID")
+	envGithubAppInstallationName = os.Getenv("GITHUB_APP_INSTALLATION_NAME")
 	envGithubWorkspaceIdentifier = os.Getenv("GITHUB_WORKSPACE_IDENTIFIER")
 	envGithubWorkspaceBranch = os.Getenv("GITHUB_WORKSPACE_BRANCH")
 	envTFEUser1 = os.Getenv("TFE_USER1")
@@ -386,6 +395,8 @@ var envGithubPolicySetBranch string
 var envGithubPolicySetPath string
 var envGithubRegistryModuleIdentifer string
 var envGithubToken string
+var envGithubAppInstallationID string
+var envGithubAppInstallationName string
 var envGithubWorkspaceIdentifier string
 var envGithubWorkspaceBranch string
 var envTFEUser1 string